+++ 3 - Largest Prime Factor/src/mathOperation/FactorOf.java
@@ -5,18 +5,18 @@ import java.util.ArrayList;
 
 public class FactorOf {
 
-	private int number;
-	private ArrayList<Integer> factorList;
+	private long number;
+	private ArrayList<Long> factorList;
 	private boolean isPrime_;
 	
 	
-	public FactorOf(int number_) {
+	public FactorOf(long number_) {
 		if (number_ <= 0) {
 			throw new ArithmeticException("Number should be bigger than 0");
 		}
 		
 		number = number_;
-		factorList = new ArrayList<Integer>();
+		factorList = new ArrayList<Long>();
 		
 		calculatePrimeFactor();
 		calculateIfPrime();
@@ -24,11 +24,11 @@ public class FactorOf {
 	
 	
 	
-	public int getNumber() {
+	public long getNumber() {
 		return number;
 	}
 	
-	public ArrayList<Integer> getFactor(){
+	public ArrayList<Long> getFactor(){
 		return factorList;
 	}
 	
@@ -44,7 +44,7 @@ public class FactorOf {
 	
 	private void calculateFirstHalfOfFactor() {
 		
-		for(int i = 1; i<= Math.sqrt(number); i++ ){
+		for(long i = 1; i<= Math.sqrt(number); i++ ){
 			
 			if (number % i == 0) {
 				factorList.add(i);
@@ -59,7 +59,7 @@ public class FactorOf {
 		//First half :  1, 2
 		//Second half : 2, 4
 		
-		int middleFactor = factorList.get(factorList.size() - 1);
+		long middleFactor = factorList.get(factorList.size() - 1);
 		
 		int numberOfiteration = factorList.size();//Need to be outside the loop, because the factorList size change with each iteration
 		
@@ -67,7 +67,7 @@ public class FactorOf {
 			
 			int index = numberOfiteration-1 - i; //Use to keep the number in order
 			
-			int reciprocalFactor = number / factorList.get(index);
+			long reciprocalFactor = number / factorList.get(index);
 			
 			if (reciprocalFactor != middleFactor) {
 				factorList.add(reciprocalFactor);
