+++ 3 - Largest Prime Factor/tests/mathOperation_test/test_PrimeFactorOf.java
@@ -11,13 +11,13 @@ import mathOperation.PrimeFactor;
 
 class test_PrimeFactorOf {
 
-	ArrayList<Integer> resultFactor;
-	ArrayList<Integer> numbers;
+	ArrayList<Long> resultFactor;
+	ArrayList<Long> numbers;
 	
 	@BeforeEach
 	void setUp() {
-		numbers = new ArrayList<Integer>();
-		resultFactor = new ArrayList<Integer>();
+		numbers = new ArrayList<Long>();
+		resultFactor = new ArrayList<Long>();
 	}
 	
 	@Test
@@ -41,7 +41,7 @@ class test_PrimeFactorOf {
 	void test_number2() {
 		numbers = PrimeFactor.PrimeFactorOf(2);
 		
-		resultFactor.add(2);
+		resultFactor.add(2L);
 		
 		assertEquals(resultFactor, numbers);
 	}
@@ -50,7 +50,7 @@ class test_PrimeFactorOf {
 	void test_number4() {
 		numbers = PrimeFactor.PrimeFactorOf(2);
 		
-		resultFactor.add(2);
+		resultFactor.add(2L);
 
 		assertEquals(resultFactor, numbers);
 	}
@@ -59,8 +59,8 @@ class test_PrimeFactorOf {
 	void test_number10() {
 		numbers = PrimeFactor.PrimeFactorOf(10);
 		
-		resultFactor.add(2);
-		resultFactor.add(5);
+		resultFactor.add(2L);
+		resultFactor.add(5L);
 		
 		assertEquals(resultFactor, numbers);
 	}
@@ -69,7 +69,7 @@ class test_PrimeFactorOf {
 	void test_number13() {
 		numbers = PrimeFactor.PrimeFactorOf(13);
 		
-		resultFactor.add(13);
+		resultFactor.add(13L);
 		
 		assertEquals(resultFactor, numbers);
 	}
